app/requirements.txt
Django==5.1

django-cors-headers==4.3.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-camel-case==1.4.2
drf-spectacular==0.28.0

django-modeltranslation==0.19.12
django-environ==0.11.2
django-unfold==0.59.0
django-phonenumber-field[phonenumbers]==8.0.0
django-silk==5.3.2
django-debug-toolbar==5.2.0

django-cachalot==2.7.0
django-redis==5.4.0

channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
uvicorn==0.34.0
websockets==14.2

Pillow==11.2.1
gunicorn==23.0.0
httpx==0.28.1
openai==1.85.0
psycopg[c]==3.2.9
pygments==2.19
sentry-sdk[django]==2.26.1
django-import-export==4.3.7
openpyxl==3.1.5
